backend/requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.2
requests==2.32.3

psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
pgvector==0.3.5

numpy==1.26.4
pandas==2.2.2

python-docx==1.1.2
pypdf==4.3.1
pytesseract==0.3.13
Pillow==10.4.0

rapidfuzz==3.9.7
rank-bm25==0.2.2

openai==1.51.2
python-dotenv==1.0.1

docx2txt==0.8

docx2txt==0.8
